all uint64 should take in BigInteger and return BigInteger even though it is really using the long representation underneath

+++ src/main/java/org/lwes/DefaultEvent.java
@@ -168,8 +168,8 @@ public abstract class DefaultEvent implements Event {
         return (long[]) get(attributeName);
     }
 
-    public long[] getUInt64Array(String attributeName) {
-        return (long[]) get(attributeName);
+    public BigInteger[] getUInt64Array(String attributeName) {
+        return (BigInteger[]) get(attributeName);
     }
 
     public String[] getStringArray(String attributeName) {
